Revert "Support Windows 10 and Windows 11 at the same time"
This reverts commit 2c11b69b6d94e20ad92df06f9f296a12e471f47e.
